refactor: rename composables package to ui package
