+++ app/src/main/java/dev/anilbeesetti/nextplayer/ui/PermissionDetailView.kt
@@ -1,4 +1,4 @@
-package dev.anilbeesetti.nextplayer.composables
+package dev.anilbeesetti.nextplayer.ui
 
 import android.Manifest
 import android.content.Intent
